% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_stats.R
\name{word_stats}
\alias{word_stats}
\title{Calculate Word Statistics}
\usage{
word_stats(tokens)
}
\arguments{
\item{tokens}{A list of character vectors from \code{\link[=tokenize]{tokenize()}}.}
}
\value{
A list with the following elements:
\describe{
\item{total_words}{Total number of words}
\item{unique_words}{Number of unique words}
\item{avg_word_length}{Average word length}
\item{lexical_diversity}{Ratio of unique to total words}
}
}
\description{
Computes summary statistics about tokenized text.
}
\examples{
tokens <- tokenize("The quick brown fox jumps over the lazy dog")
word_stats(tokens)
}
\seealso{
\code{\link[=tokenize]{tokenize()}} for creating tokens, \code{\link[=count_words]{count_words()}} for
word frequencies.
}
